requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cfscrape==2.0.5
chardet==3.0.4
idna==2.8
pkg-resources==0.0.0
requests==2.22.0
soupsieve==1.9.1
urllib3==1.25.3
